1.通过layui改造表格和上传文件功能，同时实现新建文件夹的功能(弹窗) 2.优化表格的显示位置，水平居中显示

+++ src/main/java/com/xck/util/FileUtils.java
@@ -73,7 +73,9 @@ public class FileUtils {
                 path.append(pathPart[i]).append("/");
             }
         }
-        path.delete(path.length()-1, path.length());
+        if (path.length() > 0) {
+            path.delete(path.length()-1, path.length());
+        }
 
         return path.toString();
     }
